backend/requirements.txt
pipenv
flask
psycopg2-binary
sqlalchemy
flask-cors
gunicorn
wheel
virtualenv